+++ src/components/Dropdown/Dropdown.tsx
@@ -14,7 +14,8 @@ import React, {
 } from 'react'
 
 import { Direction, Position } from './index'
-import { Wrapper, ButtonWrapper } from './Wrapper'
+import { Wrapper } from './Wrapper'
+import { ButtonWrapper } from './ButtonWrapper'
 import Items, { BaseItems } from './Items'
 
 export interface DropdownExposedProps {
@@ -50,16 +51,16 @@ export interface Options
  * @param {number} [defaultActiveItem=-1] - The index of the item that is active by default. Default is -1.
  * @param {Function} [onItemSelect] - A function that is called when an item is selected.
  *
- * Theme CSS Variables:
+ * **Theme CSS variables:**
  *
- * * --theme-dropdown-background-color
- * * --theme-dropdown-border-color
- * * --theme-dropdown-box-shadow
+ * - `--theme-dropdown-background-color`
+ * - `--theme-dropdown-border-color`
+ * - `--theme-dropdown-box-shadow`
  *
- * Base CSS Variables:
+ * **Base CSS variables:**
  *
- * * --base-dropdown-border-radius
- * * --base-dropdown-animation-time
+ * - `--base-dropdown-border-radius`
+ * - `--base-dropdown-animation-time`
  */
 const Dropdown: FC<Options> = forwardRef<DropdownExposedProps, Omit<Options, 'ref'>>(
   (
