+++ src/components/Dropdown/Item.tsx
@@ -7,22 +7,24 @@ export interface Props {
 }
 
 /**
- * Dropdown item component
+ * @name Item
+ *
+ * @description Dropdown item component
  *
  * @param {boolean} [disabled=false] - Disables the dropdown item. Default is false.
  * @param {ActionStates} [$state=undefined] - The state of the item. Defaults to undefined.
  *
- * Theme CSS variables:
+ * **Theme CSS variables:**
  *
- * * --theme-dropdown-item-background-color: Dropdown item background color.
- * * --theme-dropdown-item-border-color: Dropdown item border color.
- * * --theme-dropdown-item-color: Dropdown item text color.
- * * --theme-dropdown-item-background-color-hover: Dropdown item background color on hover.
- * * --theme-dropdown-item-border-color-hover: Dropdown item border color on hover.
- * * --theme-dropdown-item-color-hover: Dropdown item text color on hover.
- * * --theme-dropdown-item-background-color-active: Dropdown item background color when active.
- * * --theme-dropdown-item-border-color-active: Dropdown item border color when active.
- * * --theme-dropdown-item-color-active: Dropdown item text color when active.
+ * - `--theme-dropdown-item-background-color`
+ * - `--theme-dropdown-item-border-color`
+ * - `--theme-dropdown-item-color`
+ * - `--theme-dropdown-item-background-color-hover`
+ * - `--theme-dropdown-item-border-color-hover`
+ * - `--theme-dropdown-item-color-hover`
+ * - `--theme-dropdown-item-background-color-active`
+ * - `--theme-dropdown-item-border-color-active`
+ * - `--theme-dropdown-item-color-active`
  */
 const Item = styled.div.attrs<Props>(
   ({ tabIndex = 0, className = 'dbuitkDropdownItem', disabled = false }) => ({
