+++ src/components/Dropdown/Wrapper.tsx
@@ -1,5 +1,10 @@
 import { styled } from 'styled-components'
 
+/**
+ * @name Wrapper
+ *
+ * @description Dropdown main wrapper component
+ */
 export const Wrapper = styled.div.attrs<{
   $isOpen: boolean
   disabled?: boolean
@@ -11,10 +16,6 @@ export const Wrapper = styled.div.attrs<{
   position: relative;
   z-index: 0;
 
-  &.fullWidth {
-    width: 100%;
-  }
-
   &.isOpen {
     z-index: 50;
   }
@@ -25,16 +26,4 @@ export const Wrapper = styled.div.attrs<{
     opacity: 0.5;
     pointer-events: none;
   }
-`
-
-export const ButtonWrapper = styled.div.attrs(({ className = 'dbuitkDropdownButton' }) => ({
-  className,
-}))`
-  background-color: transparent;
-  border: none;
-  display: flex;
-  outline: none;
-  padding: 0;
-  user-select: none;
-  width: fit-content;
 `
