+++ src/components/SwitchThemeButton/index.tsx
@@ -124,7 +124,9 @@ const DarkIcon = styled(Dark)`
 interface Props extends ButtonHTMLAttributes<HTMLButtonElement> {}
 
 /**
- * SwitchThemeButton component - A button that switches between light and dark themes.
+ * @name SwitchThemeButton for dAppBooster
+ *
+ * @description A button that switches between light and dark themes.
  *
  * @param {MouseEventHandler<HTMLButtonElement>} onClick - Function that switches the theme.
  */
